+++ src/main/java/dev/lambdaurora/lambdabettergrass/metadata/LBGLayer.java
@@ -28,7 +28,7 @@ import java.util.function.Function;
  * Represents a layer.
  *
  * @author LambdAurora
- * @version 1.2.0
+ * @version 1.3.0
  * @since 1.0.0
  */
 public class LBGLayer {